+++ autobahn/wamp/protocol.py
@@ -67,13 +67,14 @@ class BaseSession(ObservableMixin):
 
     This class implements :class:`autobahn.wamp.interfaces.ISession`.
     """
-
-    log = txaio.make_logger()
+    log = None
 
     def __init__(self):
         """
 
         """
+        self.log = txaio.make_logger()
+
         self.set_valid_events(
             valid_events=[
                 'join',         # right before onJoin runs